+++ app/ExternalFiles/Proxy.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * This file contains the proxy-handling.
+ * This file contains the proxy tasks.
  *
  * @package external-files-in-media-library
  */
@@ -130,14 +130,14 @@ class Proxy {
 		// get the query-value.
 		$title = get_query_var( $this->get_slug() );
 
-		// get basename from request for sized images.
-		$size = array();
+		// get basename from request for sized images depending on its dimensions.
+		$dimensions = array();
 		if ( 1 === preg_match( '/(.*)-(.*)x(.*)\.(.*)/', $title, $matches ) ) {
-			$size  = array(
+			$dimensions = array(
 				absint( $matches[2] ),
 				absint( $matches[3] ),
 			);
-			$title = $matches[1] . '.' . $matches[4];
+			$title      = $matches[1] . '.' . $matches[4];
 		}
 
 		// log this event.
@@ -167,7 +167,7 @@ class Proxy {
 		}
 
 		// get cached file path.
-		$cached_file_path = $external_file_obj->get_cache_file( $size );
+		$cached_file_path = $external_file_obj->get_cache_file( $dimensions );
 
 		// bail if file does not exist.
 		if ( ! file_exists( $cached_file_path ) ) {
@@ -181,7 +181,7 @@ class Proxy {
 
 		// get the object of this file type.
 		$file_type_obj = File_Types::get_instance()->get_type_object_for_file_obj( $external_file_obj );
-		$file_type_obj->set_size( $size );
+		$file_type_obj->set_dimensions( $dimensions );
 
 		// output the proxied file.
 		$file_type_obj->get_proxied_file();
@@ -370,7 +370,7 @@ class Proxy {
 	 */
 	public function prevent_proxied_url( bool $result, File $external_file_object ): bool {
 		// bail if result is already false.
-		if( ! $result ) {
+		if ( ! $result ) {
 			return false;
 		}